infrastructure/cloudformation/pyproject.toml
[tool.poetry]
name = "website-react-pf-cdk"
version = "1.0.0"
description = "website-react-pf-cdk"
authors = [
    "Preston Frazier <preston@prestonfrazier.net>"
]
packages = [
   { include = "src" }
]

[tool.poetry.dependencies]
python = "~3.13"
boto3 = "1.42.45"
aws-cdk-lib = "2.238.0"
constructs = "10.4.5"
python-json-logger = "3.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
pytest-cov = "6.0.0"
coverage = {extras = ["toml"], version = "7.6.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v --cov --cov-report term --cov-report xml"
log_cli = true
log_cli_level = "WARNING"
testpaths = [
    "lambda_functions/*/tests"
]

[tool.coverage.run]
branch = true
include = [
    "lambda_functions/*/src/*"
]
omit = [
    "*/tests/*",
    "*__init.__*"
]

[tool.coverage.xml]
output = "coverage.xml"
